export diff viewer component
